requirements.txt
flask
python-dotenv
openai
aiohttp
requests
supabase
google-cloud-documentai
google-cloud-aiplatform
google-cloud-storage
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-genai
pydantic
asyncio
aiofiles
python-multipart
structlog
tenacity